requirements.txt
torch
torchvision
flask
flask-cors
pillow
numpy
opencv-python-headless
pytorch-grad-cam
